requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
alabaster==1.0.0
alembic==1.14.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
asttokens==3.0.0
attr==0.3.2
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
boltons==24.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
configparser==7.1.0
contextlib2==21.6.0
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.11
datasets==2.21.0
decorator==5.1.1
deep-translator==1.11.4
dill==0.3.8
dl==0.1.0
docopt==0.6.2
docutils==0.21.2
einops==0.8.0
executing==2.1.0
ffmpeg==1.4
ffpyplayer==4.5.2
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.6.1
funcsigs==1.0.2
greenlet==3.1.1
HTMLParser==0.0.2
huggingface-hub==0.26.3
HyperPyYAML==1.2.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
imagesize==1.4.1
importlib_metadata==8.5.0
ipython==8.12.3
ipywidgets==8.1.5
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
julius==0.2.7
jupyterlab_widgets==3.0.13
keyring==25.5.0
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.9
llvmlite==0.43.0
lockfile==0.12.2
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==5.1.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numba==0.60.0
numpy==1.24.1
omegaconf==2.3.0
optuna==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==18.1.0
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytorch-metric-learning==2.7.0
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
railroad==0.5.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
speechbrain==1.0.2
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tk==0.1.0
tkVideo==0.1
tokenizers==0.19.1
toolz==1.0.0
#torch==2.5.1+cu124
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1+cu124
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.2
trove-classifiers==2024.10.21.16
truststore==0.10.0
typer==0.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
urllib3-secure-extra==0.1.0
wcwidth==0.2.13
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0
